requirements.txt
# Lint
flake8==3.5.0

# Django
Django==2.0.1
django-environ==0.4.4

# Postgres Driver
psycopg2==2.7.3.2

# Wagtail
wagtail==2.0rc1

# Static and media files
django-storages==1.6.5
django-webpack-loader==0.5.0

# Logs
watchtower==0.5.2
